save work

+++ basic_types.hxx
@@ -1,7 +1,7 @@
 #ifndef _BASIC_TYPES_HXX_
 #define _BASIC_TYPES_HXX_
 
-typedef unsigned long ulong;
+typedef unsigned long long ulong;
 typedef void* any_ptr;
 
 #endif // _BASIC_TYPES_HXX_
